Refactors item picker file name.
